% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_custom2.R
\name{annotation_custom2}
\alias{annotation_custom2}
\title{Inset plots}
\usage{
annotation_custom2(
  grob,
  xmin = -Inf,
  xmax = Inf,
  ymin = -Inf,
  ymax = Inf,
  data
)
}
\arguments{
\item{grob}{The grob}

\item{xmin, xmax, ymin, ymax}{Inset plot coordinates}

\item{data}{The underlying data

Code for inset plotting from Clare West at https://www.blopig.com/blog/2019/08/combining-inset-plots-with-facets-using-ggplot2/}
}
\description{
Plot inset figures on another grob
}
