% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify}
\alias{classify}
\title{Classification analysis}
\usage{
classify(
  data,
  variables,
  grouping,
  nesting = NULL,
  method = "SVM",
  k = 5,
  nrep = 1,
  minsize = 5,
  seed = NULL,
  importance = FALSE,
  return_machine = FALSE,
  verbose = TRUE,
  pb = TRUE,
  digest = TRUE,
  test = TRUE,
  to_pcomp = NULL,
  center = TRUE,
  scale = TRUE
)
}
\arguments{
\item{data}{A data frame}

\item{variables}{The variables used to classify}

\item{grouping}{Name of the grouping variable (the labels)}

\item{nesting}{Optional nesting variable, if the analysis must be conducted separately on different subsets of the data}

\item{method}{The data mining model used. Currently supports "SVM" and "LDA".}

\item{k}{Number of bins for the k-fold cross-validation procedure}

\item{nrep}{Number of replicate analyses (i.e. number of k-fold cross validations)}

\item{minsize}{Minimum size required per group for a training data set}

\item{seed}{Optional random seed to reset at the beginning}

\item{importance}{Whether to perform sensitivity analysis on the input (takes a while)}

\item{return_machine}{Whether to return the machines (takes space)}

\item{verbose}{Whether to display messages}

\item{pb}{Whether to display progress bars}

\item{digest}{Whether to return the results in a summarized format}

\item{test}{Whether to test the results against random guessing using a binomial distribution}

\item{to_pcomp}{Variable to perform PCA on}

\item{center}{Center the PCA}

\item{scale}{Scale the PCA}

\item{add_signif}{Whether to add significance asterisk labels in an extra column}
}
\value{
A list containing the results. If \code{digest} is \code{FALSE}, a nested list on two levels with, for each replicate (first level) cross-validation bin (second level), i.e. for each machine, the following fields:
\itemize{
\item{\code{conf}}{ The confusion matrix resulting from testing the machine}
\item{\code{imp}}{ A vector of relative importance of input variable}
\item{\code{machine}}{ The machine itself}
}
or a list of such nested lists if \code{nesting} is provided, with one nested list per subset of the data.
If \code{digest} is TRUE, returns a list with the following fields:
\itemize{
\item{\code{mean}}{ A data frame with a column "accu" containing the mean accuracy across all machines on a single row, or accuracies over subsets of the data on multiple rows if \code{nesting} is provided. If \code{test} is \code{TRUE}, extra columns are added, containing the number of observations, "n", the proportion of observations used in testing the machines, "ptest", the number of observations in a testing set, "ntest", the binomial P-value assessing the significance of the obtained result, "pvalue".}
\item{\code{avg}}{ The average over replicates of the average confusion matrices over cross-validation bins, or a list of those if \code{nesting} is provided}
\item{\code{accu}}{ A vector of classification accuracy across all machines, or a data frame with a column for the nesting variable if \code{nesting} is provided}
\item{\code{confs}}{ A nested list on two levels of confusion matrices, per replicate and cross-validation bin, or a list of such nested lists if \code{nesting} is provided}
\item{\code{imp}}{ A data frame with the relative importance of each input variable for each machine, with an extra column with a nesting variable if \code{nesting} is provided}
}
}
\description{
Perform a replicated classification analysis of a multivariate dataset into categorical labels using machine learning tools and k-fold cross validation
}
