% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify2.R
\name{classify2}
\alias{classify2}
\title{Classification analysis}
\usage{
classify2(
  data,
  variables,
  grouping,
  nesting = NULL,
  method = "SVM",
  k = 5,
  nrep = 1,
  minsize = 5,
  seed = NULL,
  importance = FALSE,
  return_machine = FALSE,
  verbose = TRUE,
  pb = TRUE,
  to_pcomp = NULL,
  center = TRUE,
  scale = TRUE
)
}
\arguments{
\item{data}{A data frame}

\item{variables}{The variables used to classify}

\item{grouping}{Name of the grouping variable (the labels)}

\item{nesting}{Optional nesting variable, if the analysis must be conducted separately on different subsets of the data}

\item{method}{The data mining model used. Currently supports "SVM" and "LDA".}

\item{k}{Number of bins for the k-fold cross-validation procedure}

\item{nrep}{Number of replicate analyses (i.e. number of k-fold cross validations)}

\item{minsize}{Minimum size required per group for a training data set}

\item{seed}{Optional random seed to reset at the beginning}

\item{importance}{Whether to perform sensitivity analysis on the input (takes a while)}

\item{return_machine}{Whether to return the machines (takes space)}

\item{verbose}{Whether to display messages}

\item{pb}{Whether to display progress bars}

\item{to_pcomp}{Variable to perform PCA on}

\item{center}{Center the PCA}

\item{scale}{Scale the PCA}
}
\value{
A list of two tibbles.

The first tibble, \code{full}, contains information about each fitted machine in the analysis, including its nesting level \code{nesting}, replicate number \code{repl} and cross-validation bin \code{kbin} but also classification testing outputs such as its confusion matrix \code{confmat}, its prediction \code{accuracy} and the number of observations in the testing set \code{ntested}. If \code{return_machine} is TRUE, an extra column is added, that is a list of the actual fitted machines. If \code{importance} is TRUE, extra columns are added, named after the input variables and containing their respective relative importance scores (these are computed using the \code{rminer::Importance} function for one-dimensional sensitivity analysis).

The second tibble, \code{digested}, is a summary with, for each \code{nesting} level: the average confusion matrix \code{confmat} over all \code{nrep} replicate total confusion matrices (which are sums of \code{k} confusion matrices, one per cross-validation bin), the \code{mean} accuracy and its standard error \code{stderr} across replicates, Stouffer's Z-statistic \code{zs} used to combine P-values from binomial tests for each replicate (using the Z-transform method, see Whitlock 2005) and the corresponding combine P-value \code{pcombined}.
}
\description{
Perform a replicated classification analysis of a multivariate dataset into categorical labels using machine learning tools and k-fold cross validation
}
