% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrep.R
\name{mrep}
\alias{mrep}
\title{Repeat multiple things multiple times}
\usage{
mrep(x, n)
}
\arguments{
\item{x}{What to repeat}

\item{n}{How many times each?}
}
\description{
Repeat multiple things multiple times
}
