% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanova.R
\name{nanova}
\alias{nanova}
\title{Nested ANOVA}
\usage{
nanova(
  data,
  variables,
  grouping,
  nesting = NULL,
  univariate = TRUE,
  posthoc = TRUE,
  to_pcomp = NULL,
  center = TRUE,
  scale = TRUE,
  test = "Pillai",
  add_signif = TRUE,
  parametric = TRUE,
  seed = NULL,
  iter = 1000,
  random = NULL
)
}
\arguments{
\item{data}{A data frame}

\item{variables}{The variables to analyze}

\item{grouping}{The name of the grouping factor}

\item{nesting}{The name of the nesting factor}

\item{univariate}{Whether to perform ANOVA (TRUE) or MANOVA (FALSE)}

\item{posthoc}{Whether to perform post-hoc tests}

\item{to_pcomp}{Optional variables to perform PCA on}

\item{center, scale}{Parameters for \code{npcomp}}

\item{test}{Test to use for the MANOVA (defaults to Pillai's trace). See \code{?summary.manova}}

\item{add_signif}{Whether to add significance asterisk labels in an extra column}

\item{parametric}{Whether to perform parametric tests (linear models instead of Kruskal-Wallis if \code{univariate} is TRUE, parametric instead of semi-parametric MANOVA if \code{univariate} is FALSE)}

\item{seed}{Optional seed for the semi-parametric MANOVA}

\item{iter}{Number of iterations for the parametric bootstrapping of the semi-parametric MANOVA (defaults to the recommended number 1,000)}

\item{random}{Optional random effect for fitting a mixed model in univariate ANOVA}
}
\value{
If \code{manova} is FALSE, data frame containing the results of the ANOVAs for each variable and each subset of the data, including:

\itemize{
\item{best}{ The best-fitting model (OLS or GLS)}
\item{df_model}{ Model degrees of freedom}
\item{AICc}{ AICc of the model}
\item{dAICc}{ Difference in AICc between the GLS-model and the OLS-model}
\item{df_LRT}{ Degrees of freedom of the likelihood ratio test}
\item{loglik}{ Log-likelihood of the full model in LRT}
\item{lratio}{ Chi-square value i.e. likelihood ratio between the full and the null model in LRT}
\item{pvalue}{ P-value of the LRT}
\item{posthoc_test}{ Post-hoc test used: Wilcoxon or Tukey}
\item{posthoc_p...}{ Extra columns with the P-values for each contrast in the multiple post-hoc comparisons}
}

Otherwise, a data frame with MANOVA results for each subset, including, if \code{semiparametric} is FALSE,

\itemize{
\item{df}{ The degrees of freedom of the MANOVA}
\item{"statistic's name"}{ One of the statistics defined in \code{summary.manova} (Pillai, Wilks, etc.)}
\item{num_df}{ Numerator degrees of freedom of the F-test}
\item{denom_df}{ Denominator degrees of freedom of the F-test}
\item{pseudoF}{ Approximate F-statistic}
\item{pvalue}{ P-value of the F-test}
}

otherwise, the results of a semi-parametric MANOVA, with:

\itemize{
\item{term}{ The term being tested, probably the grouping variable}
\item{MATS}{ The modified ANOVA-type statistic value}
\item{pvalue}{ P-value computed from the parametric bootstrap resampling}
}
}
\description{
Performs (M)ANOVAs for multiple variables across different subsets of the data. Models fitted with OLS and GLS (allowing one independent residual variance per group) are compared with AIC and the best-fitting model is retained. Multiple post-hoc comparisons can be performed too.
}
\details{
The analysis of variance is performed using a likelihood ratio test between a model including the factor of interest and a null model with intercept only. The LRT is done with models fitted with maximum likelihood. Model comparison between OLS and GLS is done with models fitted with restricted maximum likelihood, that include the factor to be tested (as per Zuur et al. 2009).
}
