% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npcomp.R
\name{npcomp}
\alias{npcomp}
\title{Nested principal component analysis}
\usage{
npcomp(
  data,
  variables,
  center = TRUE,
  scale = TRUE,
  nesting = NULL,
  combine = TRUE,
  reduce = NULL
)
}
\arguments{
\item{data}{A data frame}

\item{variables}{The variables to analyze}

\item{center}{Whether to center the data}

\item{scale}{Whether to scale the data}

\item{nesting}{Optional nesting factor indicating the subsets. Defaults to NULL, where the output is that of a regular call to \code{prcomp}}

\item{combine}{If \code{nesting} is provided, whether to combine the results for the different subsets.}

\item{reduce}{If \code{combine} is TRUE, what principal components to retain from across the subsets.}
}
\value{
If \code{nesting} is NULL, the same as \code{prcomp}, otherwise, if \code{combine} is FALSE, a nested list of \code{prcomp} outputs, one for each subset. If \code{combine} is TRUE, a list of three data frames:
\itemize{
\item{\code{sdev}}{Contains the standard deviations along each principal component for each subset}
\item{\code{rotation}}{Contains the rotation matrices for each subset, bound by row}
\item{\code{x}}{The scores of each observation on each principal component}
}
}
\description{
Perform principal component analyses across multiple subsets of the data
}
