% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nspcortest.R
\name{nspcortest}
\alias{nspcortest}
\title{Nest spatial autocorrelation}
\usage{
nspcortest(
  data,
  variables,
  nesting = NULL,
  nperm = 1000,
  seed = NULL,
  lon = "longitude",
  lat = "latitude",
  verbose = TRUE,
  pb = TRUE,
  to_pcomp = NULL,
  center = TRUE,
  scale = TRUE
)
}
\arguments{
\item{data}{A data frame}

\item{variables}{The variables to analyze}

\item{nesting}{Optional nesting factor}

\item{nperm}{Number of permutations}

\item{seed}{Optional random seed}

\item{lon, lat}{Longitude and latitude columns}

\item{verbose}{Whether to display messages}

\item{pb}{Whether to display progress bars}

\item{to_pcomp}{Variable to perform PCA on}

\item{center, scale}{Parameters for \code{npcomp}}
}
\value{
A list with two data frames: one with the results of the correlation test on each subset, including observed Pearson's correlation, P-values and number of sites, and the second with the geographical coordinates and means of each trait for every site in the data.
}
\description{
Perform multiple spatial autocorrelation tests within subsets of the data (e.g. islands) using permutations of individual observations. Particularly suited to designs with few samping sites per subset.
}
