% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_classif.R
\name{plot_classif}
\alias{plot_classif}
\title{Plot classification results}
\usage{
plot_classif(
  res,
  facets = NULL,
  bins = 15,
  fill = "seagreen",
  alpha = 0.5,
  ylim = c(0, 200),
  norm = 2,
  low = "white",
  high = "darkgreen",
  limits = c(0, 1),
  xmin = 0,
  xmax = 0.35,
  ymin = 120,
  ymax = 200,
  dfac = 500,
  rounding = 4,
  signif = 0.05,
  px = 1,
  py = 190,
  phjust = 1,
  type = "histogram",
  add_insets = TRUE,
  add_null = TRUE,
  add_pvalues = TRUE
)
}
\arguments{
\item{res}{The output of \code{classify}}

\item{facets}{Optional name of the facetting (nesting) variable}

\item{bins}{Number of bins for the histograms}

\item{fill}{Color of the histograms}

\item{alpha}{Transparency of the histograms}

\item{ylim}{Limits of the vertical axis}

\item{norm}{How to normalize the average confusion matrix: 0 for no normalization, 1 for by row, and 2 for by column}

\item{low, high, limits}{Parameters for the gradient fill scale in plotting the confusion matrices}

\item{xmin, xmax, ymin, ymax}{Coordinates of the confusion matrices as insets in the histogram plots}

\item{dfac}{Factor to scale the height of the null density line}

\item{rounding}{How to round the P-values}

\item{signif}{Significance threshold for P-values}

\item{px, py, phjust}{Parameters for \code{geom_text} when displaying P-values}

\item{type}{Type of plot to return: "histogram" (for histograms of classification success) or "confusion" (confusion matrix)}

\item{add_insets}{Whether to add confusion matrices as insets in the histogram plots}

\item{add_null}{Whether to add null density lines in the histogram plots}

\item{add_pvalues}{Whether to add P-values in the histogram plots}
}
\value{
A ggplot


}
\description{
Plot the output of the \code{classify} function
}
