% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_covariance.R
\name{test_covariance}
\alias{test_covariance}
\title{Test homogeneity of covariance matrices}
\usage{
test_covariance(d, variables, grouping, nesting = NULL)
}
\arguments{
\item{d}{A data frame}

\item{variables}{What variables to test?}

\item{grouping}{Grouping variable}

\item{nesting}{Nesting variable}
}
\value{
A data frame with the chi-square statistics, degrees of freedom and P-value of Box's M-test performed between groups within each subset
}
\description{
Uses Box's M-test to assess the homogeneity of multivariate covariance matrices between groups within each subset
}
