% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_multinorm.R
\name{test_multinorm}
\alias{test_multinorm}
\title{Test multivariate normality}
\usage{
test_multinorm(
  d,
  variables,
  grouping,
  nesting = NULL,
  univariate = FALSE,
  add_signif = TRUE
)
}
\arguments{
\item{d}{A data frame}

\item{variables}{What variables to test?}

\item{grouping}{Grouping variable}

\item{nesting}{Nesting variable}

\item{univariate}{Special option to return the results of multiple univariate Sharpiro tests}

\item{add_signif}{Whether to add significance asterisk labels in an extra column}
}
\value{
A data frame with HZ statistics and P-values for each group in each subset, or a data frame with W statistics and P-values for each variable in each group and each subset, if \code{univariate} is TRUE.
}
\description{
Uses Henze-Zirkler's (or Shapiro's) test to assess the multivariate normality across groups and subsets of the data
}
