% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_outliers.R
\name{test_outliers}
\alias{test_outliers}
\title{Test multivariate outliers}
\usage{
test_outliers(d, variables, grouping, nesting = NULL)
}
\arguments{
\item{d}{A data frame}

\item{variables}{What variables to test?}

\item{grouping}{Grouping variable}

\item{nesting}{Nesting variable}
}
\value{
A nested list of outlier identities across groups and subsets
}
\description{
Finds multivariate outliers across groups and subsets using a quantile method based on Mahalanobis distance
}
