% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_signif.R
\name{add_signif}
\alias{add_signif}
\title{Add significance column}
\usage{
add_signif(
  df,
  pcol = "pvalue",
  thresholds = c(0.05, 0.01, 0.001),
  labels = c("*", "**", "***"),
  colname = ""
)
}
\arguments{
\item{df}{A data frame}

\item{pcol}{Name of the column containing P-values}

\item{thresholds}{A vector of P-value thresholds, in decreasing order}

\item{labels}{A vector of labels, one for each threshold}

\item{colname}{Optional name of the significance label column. Defaults to none}
}
\value{
A data frame
}
\description{
Adds a column with significance labels to a table of statistical results.
}
