% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanova.R
\name{nanova}
\alias{nanova}
\title{Nested ANOVA}
\usage{
nanova(
  data,
  variables,
  grouping,
  nesting = NULL,
  posthoc = TRUE,
  to_pcomp = NULL,
  center = TRUE,
  scale = TRUE,
  manova = FALSE,
  test = "Pillai",
  assumptions = TRUE,
  univariate = FALSE,
  kw = FALSE
)
}
\arguments{
\item{data}{A data frame}

\item{variables}{The variables to analyze}

\item{grouping}{The name of the grouping factor}

\item{nesting}{The name of the nesting factor}

\item{posthoc}{Whether to perform post-hoc tests (if the GLS-model is the best fit, nonparametetric Wilcoxonn tests are used for post-hoc comparisons, otherwise Tukey's HSD test is used)}

\item{to_pcomp}{Optional variables to perform PCA on}

\item{center, scale}{Parameters for \code{npcomp}}

\item{manova}{Whether to perform a MANOVA instead of multiple ANOVAs within each subset}

\item{test}{Test to use for the MANOVA (defaults to Pillai's trace). See \code{?summary.manova}}

\item{assumptions}{Whether to perform tests of assumptions for the MANOVA}

\item{univariate}{Whether to test for multiple univariate normality instead of multivariate normality. See \code{?test_multinorm}}

\item{kw}{Whether to perform Kruskal-Wallis tests instead of ANOVAs}

\item{add_signif}{Whether to add significance asterisk labels in an extra column}
}
\value{
If \code{manova} is FALSE, data frame containing the results of the ANOVAs for each variable and each subset of the data, including:

\itemize{
\item{best_fit}{ The best-fitting model (OLS or GLS)}
\item{df_model}{ Model degrees of freedom}
\item{AICc}{ AICc of the model}
\item{dAICc}{ Difference in AICc between the GLS-model and the OLS-model}
\item{df_LRT}{ Degrees of freedom of the likelihood ratio test}
\item{loglik}{ Log-likelihood of the full model in LRT}
\item{lratio}{ Chi-square value i.e. likelihood ratio between the full and the null model in LRT}
\item{pvalue}{ P-value of the LRT}
\item{posthoc_test}{ Post-hoc test used: Wilcoxon or Tukey}
\item{posthoc_p...}{ Extra columns with the P-values for each contrast in the multiple post-hoc comparisons}
}

Otherwise, a data frame with MANOVA results for each subset, including:

\itemize{
\item{df}{ The degrees of freedom of the MANOVA}
\item{"statistic's name"}{ One of the statistics defined in \code{summary.manova} (Pillai, Wilks, etc.)}
\item{num_df}{ Numerator degrees of freedom of the F-test}
\item{denom_df}{ Denominator degrees of freedom of the F-test}
\item{pseudoF}{ Approximate F-statistic}
\item{pvalue}{ P-value of the F-test}
}

If \code{assumptions} is TRUE, the resulting data frame is combined in a list with a list of three elements:
\itemize{
\item{multinorm}{ A data frame testing multivariate normality within each group and each subset. See \code{?test_multinorm}}
\item{cov}{ A data frame testing the homogeneity of covariance matrices between groups within each subset. See \code{?test_covaraince}}
\item{outliers}{ A nested list of outlier observations within each group and each subset. See \code{?test_outliers}}
}
}
\description{
Performs (M)ANOVAs for multiple variables across different subsets of the data. Models fitted with OLS and GLS (allowing one independent residual variance per group) are compared with AIC and the best-fitting model is retained. Multiple post-hoc comparisons can be performed too.
}
\details{
The analysis of variance is performed using a likelihood ratio test between a model including the factor of interest and a null model with intercept only. The LRT is done with models fitted with maximum likelihood. Model comparison between OLS and GLS is done with models fitted with restricted maximum likelihood, that include the factor to be tested (as per Zuur et al. 2009).
}
